chore(pmu): Rename edit to details
