+++ app/population/controllers/edit/details.test.js
@@ -3,12 +3,12 @@ const mockPopulationService = {
   update: sinon.stub(),
 }
 
-const FormWizardController = require('../../../common/controllers/form-wizard')
+const FormWizardController = require('../../../../common/controllers/form-wizard')
 
-const Controller = require('./edit')
+const Controller = require('./details')
 
 describe('Population controllers', function () {
-  describe('#edit()', function () {
+  describe('details controller', function () {
     let req
     let res
     let next
@@ -42,9 +42,11 @@ describe('Population controllers', function () {
         services: {
           population: mockPopulationService,
         },
+        t: sinon.fake(),
       }
       res = {
         redirect: sinon.fake(),
+        breadcrumb: sinon.fake(),
         locals: {},
       }
 
@@ -221,6 +223,39 @@ describe('Population controllers', function () {
           expect(req.journeyModel.reset).not.to.have.been.called
           expect(req.sessionModel.reset).not.to.have.been.called
           expect(next).to.have.been.calledWith(error)
+        })
+      })
+    })
+
+    describe('setBreadcrumbs', function () {
+      context('with an existing population', function () {
+        beforeEach(function () {
+          req.population = {}
+          controllerInstance.setBreadcrumbs(req, res, next)
+        })
+        it('should use update page title', function () {
+          expect(req.t).to.have.been.calledWith(
+            'population::edit.page_title_update'
+          )
+        })
+
+        it('should call next', function () {
+          expect(next).to.have.been.calledWith()
+        })
+      })
+
+      context('with an new population', function () {
+        beforeEach(async function () {
+          await controllerInstance.setBreadcrumbs(req, res, next)
+        })
+        it('should use new page title', function () {
+          expect(req.t).to.have.been.calledWith(
+            'population::edit.page_title_new'
+          )
+        })
+
+        it('should call next', function () {
+          expect(next).to.have.been.calledWith()
         })
       })
     })
